Combined MovieDisplay and ShowDisplay
